backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
httpx==0.27.0
beautifulsoup4==4.12.3
python-whois==0.8.0
dnspython==2.6.1
python-Levenshtein==0.25.1
pydantic==2.8.2
rich==13.7.1
